extract config
